test: adapt tests for uint64/uint16 support
Update the following test files to use uint64 for epoch/block parameters and uint16 for BPS fields:
- AccountLockupSettlement.t.sol
- Fees.t.sol
- RailGetters.t.sol
- RateChangeQueue.t.sol
- AccountManagement.t.sol
- OperatorApproval.t.sol
- RailSettlement.t.sol

+++ test/AccountLockupSettlement.t.sol
@@ -46,7 +46,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             DEPOSIT_AMOUNT * 2,
             0,
             0,
-            block.number
+            uint64(block.number)
         );
     }
 
@@ -64,7 +64,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             USER2,
             OPERATOR,
             lockupRate, // payment rate
-            lockupPeriod, // lockup period
+            uint64(lockupPeriod), // lockup period
             0, // no fixed lockup
             address(0) // no fixed lockup
         );
@@ -89,7 +89,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             DEPOSIT_AMOUNT * 2,
             expectedLockup,
             lockupRate,
-            block.number
+            uint64(block.number)
         );
     }
 
@@ -134,7 +134,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             DEPOSIT_AMOUNT, // expected funds
             expectedLockup, // expected lockup
             lockupRate, // expected lockup rate
-            expectedSettlementBlock // expected settlement block
+            uint64(expectedSettlementBlock) // expected settlement block
         );
     }
 
@@ -155,7 +155,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             USER2,
             OPERATOR,
             lockupRate, // 1 token per block
-            lockupPeriod, // Lockup period of 30 blocks
+            uint64(lockupPeriod), // Lockup period of 30 blocks
             initialLockup, // initial fixed lockup of 10 ether
             address(0) // no fixed lockup
         );
@@ -177,7 +177,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             DEPOSIT_AMOUNT * 3, // expected funds
             expectedLockup, // expected lockup
             lockupRate, // expected lockup rate
-            block.number // expected settlement block
+            uint64(block.number) // expected settlement block
         );
     }
 
@@ -206,7 +206,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             DEPOSIT_AMOUNT,
             DEPOSIT_AMOUNT,
             0, // no payment rate
-            block.number
+            uint64(block.number)
         );
 
         helper.makeDeposit(USER1, USER1, 1); // Adding more funds
@@ -258,7 +258,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             USER2,
             OPERATOR,
             lockupRate, // 1 ether per block
-            lockupPeriod, // Lockup period of 10 blocks
+            uint64(lockupPeriod), // Lockup period of 10 blocks
             initialLockup, // 50 ether fixed lockup
             address(0) // no fixed lockup
         );
@@ -284,7 +284,7 @@ contract AccountLockupSettlementTest is Test, BaseTestHelper {
             60 ether, // expected funds
             60 ether, // expected lockup
             lockupRate, // expected lockup rate
-            block.number // expected settlement block
+            uint64(block.number) // expected settlement block
         );
     }
 }
